feat: add framer-motion dependency for animations, update feedback rate validation to allow values between 0 and 10, and adjust Feedback entity to use float type for rate

+++ src/entities/Feedback.ts
@@ -16,7 +16,7 @@ export class Feedback {
 	@Column({ type: "text", nullable: false })
 	response!: string;
 
-	@Column({ type: "integer", nullable: false })
+	@Column({ type: "float", nullable: false })
 	rate!: number;
 
 	@CreateDateColumn({ type: "datetime" })
